Update notes as Array of Objects in local storage
Makes use of an Array of Objects to store multiple notes instead of
originally used Object with title and detail as its key-value pair in
its prior storage as that model limited content for a unique title due
to which even multiple notes having no or blank title were not possible
because of value overwrite updated under same key in JavaScript.
Adds comments for each of App and NotesCard component.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>