Add edit and delete note functionality +UI reflect
Improves page CSS and adds functionality to edit and delete an
individual note on clicking it by prompting to update its values and
offering to save update or delete it as per emptiness of response.
Without proper buttons or UI describing it, this is initiated as an
indirect or hidden feature.
Even updated header in page <nav> and open source link in footer.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>