Add New Note with Modal and store note timestamp
Uses Bootstrap Modal to add a new note using form in page instead of
previous Browser prompt to add a new note. This Modal form also stores
note creation timestamp data with a new note creation, and closes on
non-empty submission.
Captured date field is included with NoteCard but with d-none to hide it
from view, which can be provided as a setting later.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>