Add sort notes feature, +space in Toggle Dark Mode
Adds a new option to enable sorting notes on page in one go based on
their title or timestamp with ascending and descending sort options
available in both case, totaling to 4 sorting options with two
complementing another ones on ascending-descending order pair.
Also, noted that space separating dark mode icon with text starting with
'Toggle' actually displays a space between icon and Toggle Dark Mode, so
commiting that change as well in it.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>