Navbar toggler minify, Add Note modal overflow fix
Navbar toggler minified with toggler icon <span> directly used for
toggling instead of a <button> container originally used for toggling
navbar in smaller screens such as in a mobile device. Sort notes
dropdown options to also highlight with chocolate background instead of
Bootstrap primary bluish color.
Also, sets w-auto auto width as intended with .modal-dialog class which
as combined with modal-content class overridden width from auto to 100%
which caused some overflow to the right of the modal with some margin
set in place for smaller screens.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>