Documentation: Improve ending note in README.md
Edited the last part of README file to link to issues and pull requests
tabs for the repository page (with relative link paths as /issues and
/pulls respectively from the README markdown file preview view, two
level up to eliminate /blob/main that is in use by GitHub to navigate
around a directory or file) and minor formatting update to local server
commands that have been tested for local source deployments to highlight
them at the end. Additionally, fixes one first word capitalization
following ... 'example:' label on a line as it is start of a separate
sentence, for grammatical correctness.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>